requirements.txt
opencv-python==4.8.1.78
transformers==4.40.0
torch>=2.0.0
torchvision>=0.15.0
pillow==10.1.0
numpy==1.24.3
pyyaml==6.0
